refactor: use InclusiveNodeWatchFileSystem as changes source
