+++ src/hooks/getChangedFiles.ts
@@ -1,21 +1,15 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import webpack from 'webpack';
 import path from 'path';
-import { getWatcher } from './getWatcher';
+import { CompilerWithWatchFileSystem } from '../watch/CompilerWithWatchFileSystem';
+import { InclusiveNodeWatchFileSystem } from '../watch/InclusiveNodeWatchFileSystem';
 
 function getChangedFiles(compiler: webpack.Compiler): string[] {
-  let changedFiles: string[] = [];
+  const watchFileSystem = (compiler as CompilerWithWatchFileSystem<InclusiveNodeWatchFileSystem>)
+    .watchFileSystem;
 
-  if ((compiler as any).modifiedFiles) {
-    // webpack 5+
-    changedFiles = Array.from((compiler as any).modifiedFiles);
-  } else {
-    const watcher = getWatcher(compiler);
-    // webpack 4
-    changedFiles = Object.keys((watcher && watcher.mtimes) || {});
-  }
-
-  return changedFiles.map((changedFile) => path.normalize(changedFile));
+  return watchFileSystem
+    ? Array.from(watchFileSystem.changedFiles).map((changedFile) => path.normalize(changedFile))
+    : [];
 }
 
 export { getChangedFiles };
