+++ src/hooks/getDeletedFiles.ts
@@ -1,23 +1,15 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import webpack from 'webpack';
 import path from 'path';
-import { ForkTsCheckerWebpackPluginState } from '../ForkTsCheckerWebpackPluginState';
+import { CompilerWithWatchFileSystem } from '../watch/CompilerWithWatchFileSystem';
+import { InclusiveNodeWatchFileSystem } from '../watch/InclusiveNodeWatchFileSystem';
 
-function getDeletedFiles(
-  compiler: webpack.Compiler,
-  state: ForkTsCheckerWebpackPluginState
-): string[] {
-  let deletedFiles: string[] = [];
+function getDeletedFiles(compiler: webpack.Compiler): string[] {
+  const watchFileSystem = (compiler as CompilerWithWatchFileSystem<InclusiveNodeWatchFileSystem>)
+    .watchFileSystem;
 
-  if ((compiler as any).removedFiles) {
-    // webpack 5+
-    deletedFiles = Array.from((compiler as any).removedFiles || []);
-  } else {
-    // webpack 4
-    deletedFiles = [...state.removedFiles];
-  }
-
-  return deletedFiles.map((changedFile) => path.normalize(changedFile));
+  return watchFileSystem
+    ? Array.from(watchFileSystem.removedFiles).map((removeFile) => path.normalize(removeFile))
+    : [];
 }
 
 export { getDeletedFiles };
