+++ src/watch/CompilerWithWatchFileSystem.ts
@@ -1,0 +1,9 @@
+import webpack from 'webpack';
+import { WatchFileSystem } from './WatchFileSystem';
+
+interface CompilerWithWatchFileSystem<TWatchFileSystem extends WatchFileSystem = WatchFileSystem>
+  extends webpack.Compiler {
+  watchFileSystem?: TWatchFileSystem;
+}
+
+export { CompilerWithWatchFileSystem };
